requirements.txt
flask>=3.0.0
python-dotenv>=1.0.0
gunicorn>=22.0.0
Flask-SQLAlchemy>=3.1.1
psycopg2-binary>=2.9.9